fix: Refactor session update and delete actions to use dedicated handler functions
